Refactor to move cmd_actions to module and remove global SC
