+++ hopper/__init__.py
@@ -1,1 +1,1 @@
-from .hopper import run_cmds, logger, SC
+from .hopper import run_cmds, logger
